+++ src/ExtendedXmlSerializer/ExtensionModel/References/IMultipleReferencesAllowed.cs
@@ -1,0 +1,6 @@
+namespace ExtendedXmlSerializer.ExtensionModel.References;
+
+interface IMultipleReferencesAllowed
+{
+	bool Allowed { get; }
+}
