+++ src/ExtendedXmlSerializer/ExtensionModel/References/IReferenceView.cs
@@ -1,0 +1,5 @@
+using ExtendedXmlSerializer.Core.Sources;
+
+namespace ExtendedXmlSerializer.ExtensionModel.References;
+
+interface IReferenceView : IParameterizedSource<object, ReferenceResult> {}
