+++ src/ExtendedXmlSerializer/ExtensionModel/References/ReferenceCompleted.cs
@@ -1,0 +1,8 @@
+namespace ExtendedXmlSerializer.ExtensionModel.References;
+
+sealed class ReferenceCompleted
+{
+	public static ReferenceCompleted Default { get; } = new();
+
+	ReferenceCompleted() {}
+}
